Compute tighter ibp bounds from affine ones automatically
- do it in InputsOutputsSpec.get_fullinputs_from_inputsformode()
  by default
- compute_ibp_from_affine=False deactivate it for mode=AFFINE
- tight=False deactivate it for mode=HYBRID
- to avoid circular imports issues, get_upper/get_lower() are put in
  decomon.core
- get_fullinputs_by_type_from_inputsformode_to_merge() compute also
  tight bounds by default in HYBRID case
