+++ src/decomon/layers/maxpooling.py
@@ -6,10 +6,9 @@ import tensorflow.keras.backend as K
 from tensorflow.keras.backend import conv2d
 from tensorflow.keras.layers import InputSpec, MaxPooling2D
 
-from decomon.core import ForwardMode, PerturbationDomain
+from decomon.core import ForwardMode, PerturbationDomain, get_lower, get_upper
 from decomon.layers.core import DecomonLayer
 from decomon.layers.utils import max_
-from decomon.utils import get_lower, get_upper
 
 
 # step 1: compute the maximum
