+++ src/decomon/models/backward_cloning.py
@@ -18,13 +18,14 @@ from decomon.core import (
     PerturbationDomain,
     Slope,
     get_affine,
+    get_lower,
     get_mode,
+    get_upper,
 )
 from decomon.layers.utils import softmax_to_linear as softmax_2_linear
 from decomon.models.crown import Convert2BackwardMode, Fuse, MergeWithPrevious
 from decomon.models.forward_cloning import OutputMapDict
 from decomon.models.utils import Convert2Mode, get_depth_dict
-from decomon.utils import get_lower, get_upper
 
 
 def get_disconnected_input(
