+++ src/decomon/models/utils.py
@@ -24,10 +24,12 @@ from decomon.core import (
     ForwardMode,
     InputsOutputsSpec,
     PerturbationDomain,
+    get_lower,
     get_mode,
+    get_upper,
 )
 from decomon.layers.utils import is_a_merge_layer
-from decomon.utils import get_lower, get_lower_layer, get_upper, get_upper_layer
+from decomon.utils import get_lower_layer, get_upper_layer
 
 try:
     from deel.lip.layers import LipschitzLayer
