+++ src/decomon/utils.py
@@ -8,7 +8,6 @@ from tensorflow.math import greater_equal
 from tensorflow.types.experimental import TensorLike
 
 from decomon.core import (
-    BallDomain,
     BoxDomain,
     ForwardMode,
     GridDomain,
@@ -17,6 +16,10 @@ from decomon.core import (
     Slope,
     get_affine,
     get_ibp,
+    get_lower,
+    get_lower_box,
+    get_upper,
+    get_upper_box,
 )
 
 TensorFunction = Callable[[TensorLike], tf.Tensor]
@@ -75,344 +78,6 @@ def softsign_prime(x: TensorLike) -> tf.Tensor:
     """
 
     return K.cast(1.0, dtype=x.dtype) / K.pow(K.cast(1.0, dtype=x.dtype) + K.abs(x), K.cast(2, dtype=x.dtype))
-
-
-##############
-# SYMBOLIC UPPER/ LOWER BOUNDS
-# compute symbolically constant upper and lower
-# with the current knowledge of the perturbation domain considered
-##############
-
-# case 1: a box
-def get_upper_box(x_min: tf.Tensor, x_max: tf.Tensor, w: tf.Tensor, b: tf.Tensor, **kwargs: Any) -> tf.Tensor:
-    """#compute the max of an affine function
-    within a box (hypercube) defined by its extremal corners
-
-    Args:
-        x_min: lower bound of the box domain
-        x_max: upper bound of the box domain
-        w: weights of the affine function
-        b: bias of the affine function
-
-    Returns:
-        max_(x >= x_min, x<=x_max) w*x + b
-    """
-
-    if len(w.shape) == len(b.shape):  # identity function
-        return x_max
-
-    # split into positive and negative components
-    z_value = K.cast(0.0, dtype=x_min.dtype)
-    w_pos = K.maximum(w, z_value)
-    w_neg = K.minimum(w, z_value)
-
-    x_min_out = x_min + z_value * x_min
-    x_max_out = x_max + z_value * x_max
-
-    for _ in range(len(w.shape) - len(x_max.shape)):
-        x_min_out = K.expand_dims(x_min_out, -1)
-        x_max_out = K.expand_dims(x_max_out, -1)
-
-    return K.sum(w_pos * x_max_out + w_neg * x_min_out, 1) + b
-
-
-def get_lower_box(x_min: tf.Tensor, x_max: tf.Tensor, w: tf.Tensor, b: tf.Tensor, **kwargs: Any) -> tf.Tensor:
-    """
-    Args:
-        x_min: lower bound of the box domain
-        x_max: upper bound of the box domain
-        w_l: weights of the affine lower bound
-        b_l: bias of the affine lower bound
-
-    Returns:
-        min_(x >= x_min, x<=x_max) w*x + b
-    """
-
-    if len(w.shape) == len(b.shape):
-        return x_min
-
-    z_value = K.cast(0.0, dtype=x_min.dtype)
-
-    w_pos = K.maximum(w, z_value)
-    w_neg = K.minimum(w, z_value)
-
-    x_min_out = x_min + z_value * x_min
-    x_max_out = x_max + z_value * x_max
-
-    for _ in range(len(w.shape) - len(x_max.shape)):
-        x_min_out = K.expand_dims(x_min_out, -1)
-        x_max_out = K.expand_dims(x_max_out, -1)
-
-    return K.sum(w_pos * x_min_out + w_neg * x_max_out, 1) + b
-
-
-# case 2 : a ball
-def get_lq_norm(x: tf.Tensor, p: float, axis: int = -1) -> tf.Tensor:
-    """compute Lp norm (p=1 or 2)
-
-    Args:
-        x: tensor
-        p: the power must be an integer in (1, 2)
-        axis: the axis on which we compute the norm
-
-    Returns:
-        ||w||^p
-    """
-    if p == 1:
-        x_q = K.max(K.abs(x), axis)
-    elif p == 2:
-        x_q = K.sqrt(K.sum(K.pow(x, p), axis))
-    else:
-        raise NotImplementedError("p must be equal to 1 or 2")
-
-    return x_q
-
-
-def get_upper_ball(x_0: tf.Tensor, eps: float, p: float, w: tf.Tensor, b: tf.Tensor, **kwargs: Any) -> tf.Tensor:
-    """max of an affine function over an Lp ball
-
-    Args:
-        x_0: the center of the ball
-        eps: the radius
-        p: the type of Lp norm considered
-        w: weights of the affine function
-        b: bias of the affine function
-
-    Returns:
-        max_(|x - x_0|_p<= eps) w*x + b
-    """
-    if len(w.shape) == len(b.shape):
-        return x_0 + eps
-
-    if p == np.inf:
-        # compute x_min and x_max according to eps
-        x_min = x_0 - eps
-        x_max = x_0 + eps
-        return get_upper_box(x_min, x_max, w, b)
-
-    else:
-        # use Holder's inequality p+q=1
-        # ||w||_q*eps + w*x_0 + b
-
-        if len(kwargs):
-            return get_upper_ball_finetune(x_0, eps, p, w, b, **kwargs)
-
-        upper = eps * get_lq_norm(w, p, axis=1) + b
-
-        for _ in range(len(w.shape) - len(x_0.shape)):
-            x_0 = K.expand_dims(x_0, -1)
-
-        return K.sum(w * x_0, 1) + upper
-
-
-def get_lower_ball(x_0: tf.Tensor, eps: float, p: float, w: tf.Tensor, b: tf.Tensor, **kwargs: Any) -> tf.Tensor:
-    """min of an affine fucntion over an Lp ball
-
-    Args:
-        x_0: the center of the ball
-        eps: the radius
-        p: the type of Lp norm considered
-        w: weights of the affine function
-        b: bias of the affine function
-
-    Returns:
-        min_(|x - x_0|_p<= eps) w*x + b
-    """
-    if len(w.shape) == len(b.shape):
-        return x_0 - eps
-
-    if p == np.inf:
-        # compute x_min and x_max according to eps
-        x_min = x_0 - eps
-        x_max = x_0 + eps
-        return get_lower_box(x_min, x_max, w, b)
-
-    else:
-        # use Holder's inequality p+q=1
-        # ||w||_q*eps + w*x_0 + b
-
-        if len(kwargs):
-            return get_lower_ball_finetune(x_0, eps, p, w, b, **kwargs)
-
-        lower = -eps * get_lq_norm(w, p, axis=1) + b
-
-        for _ in range(len(w.shape) - len(x_0.shape)):
-            x_0 = K.expand_dims(x_0, -1)
-
-        return K.sum(w * x_0, 1) + lower
-
-
-def get_lower_ball_finetune(
-    x_0: tf.Tensor, eps: float, p: float, w: tf.Tensor, b: tf.Tensor, **kwargs: Any
-) -> tf.Tensor:
-
-    if "finetune_lower" in kwargs and "upper" in kwargs or "lower" in kwargs:
-
-        alpha = kwargs["finetune_lower"]
-        # assume alpha is the same shape as w, minus the batch dimension
-        n_shape = len(w.shape) - 2
-
-        if "upper" and "lower" in kwargs:
-            upper = kwargs["upper"]  # flatten vector
-            lower = kwargs["lower"]  # flatten vector
-
-            upper_reshaped = np.reshape(upper, [1, -1] + [1] * n_shape)
-            lower_reshaped = np.reshape(lower, [1, -1] + [1] * n_shape)
-
-            w_alpha = w * alpha[None]
-            w_alpha_bar = w * (1 - alpha)
-
-            score_box = K.sum(K.maximum(0.0, w_alpha_bar) * lower_reshaped, 1) + K.sum(
-                K.minimum(0.0, w_alpha_bar) * upper_reshaped, 1
-            )
-            score_ball = get_lower_ball(x_0, eps, p, w_alpha, b)
-
-            return score_box + score_ball
-
-        if "upper" in kwargs:
-
-            upper = kwargs["upper"]  # flatten vector
-            upper_reshaped = np.reshape(upper, [1, -1] + [1] * n_shape)
-
-            w_alpha = K.minimum(0, w) * alpha[None] + K.maximum(0.0, w)
-            w_alpha_bar = K.minimum(0, w) * (1 - alpha[None])
-
-            score_box = K.sum(K.minimum(0.0, w_alpha_bar) * upper_reshaped, 1)
-            score_ball = get_lower_ball(x_0, eps, p, w_alpha, b)
-
-            return score_box + score_ball
-
-        if "lower" in kwargs:
-
-            lower = kwargs["lower"]  # flatten vector
-            lower_reshaped = np.reshape(lower, [1, -1] + [1] * n_shape)
-
-            w_alpha = K.maximum(0, w) * alpha[None] + K.minimum(0.0, w)
-            w_alpha_bar = K.maximum(0, w) * (1 - alpha[None])
-
-            score_box = K.sum(K.maximum(0.0, w_alpha_bar) * lower_reshaped, 1)
-            score_ball = get_lower_ball(x_0, eps, p, w_alpha, b)
-
-            return score_box + score_ball
-
-    return get_lower_ball(x_0, eps, p, w, b)
-
-
-def get_upper_ball_finetune(
-    x_0: tf.Tensor, eps: float, p: float, w: tf.Tensor, b: tf.Tensor, **kwargs: Any
-) -> tf.Tensor:
-
-    if "finetune_upper" in kwargs and "upper" in kwargs or "lower" in kwargs:
-
-        alpha = kwargs["finetune_upper"]
-        # assume alpha is the same shape as w, minus the batch dimension
-        n_shape = len(w.shape) - 2
-
-        if "upper" and "lower" in kwargs:
-            upper = kwargs["upper"]  # flatten vector
-            lower = kwargs["lower"]  # flatten vector
-
-            upper_reshaped = np.reshape(upper, [1, -1] + [1] * n_shape)
-            lower_reshaped = np.reshape(lower, [1, -1] + [1] * n_shape)
-
-            w_alpha = w * alpha[None]
-            w_alpha_bar = w * (1 - alpha)
-
-            score_box = K.sum(K.maximum(0.0, w_alpha_bar) * upper_reshaped, 1) + K.sum(
-                K.minimum(0.0, w_alpha_bar) * lower_reshaped, 1
-            )
-            score_ball = get_lower_ball(x_0, eps, p, w_alpha, b)
-
-            return score_box + score_ball
-
-        if "upper" in kwargs:
-
-            upper = kwargs["upper"]  # flatten vector
-            upper_reshaped = np.reshape(upper, [1, -1] + [1] * n_shape)
-
-            w_alpha = K.minimum(0, w) * alpha[None] + K.maximum(0.0, w)
-            w_alpha_bar = K.minimum(0, w) * (1 - alpha[None])
-
-            score_box = K.sum(K.maximum(0.0, w_alpha_bar) * upper_reshaped, 1)
-            score_ball = get_lower_ball(x_0, eps, p, w_alpha, b)
-
-            return score_box + score_ball
-
-        if "lower" in kwargs:
-
-            lower = kwargs["lower"]  # flatten vector
-            lower_reshaped = np.reshape(lower, [1, -1] + [1] * n_shape)
-
-            w_alpha = K.maximum(0, w) * alpha[None] + K.minimum(0.0, w)
-            w_alpha_bar = K.maximum(0, w) * (1 - alpha[None])
-
-            score_box = K.sum(K.minimum(0.0, w_alpha_bar) * lower_reshaped, 1)
-            score_ball = get_lower_ball(x_0, eps, p, w_alpha, b)
-
-            return score_box + score_ball
-
-    return get_upper_ball(x_0, eps, p, w, b)
-
-
-def get_upper(
-    x: tf.Tensor, w: tf.Tensor, b: tf.Tensor, perturbation_domain: Optional[PerturbationDomain] = None, **kwargs: Any
-) -> tf.Tensor:
-    """Meta function that aggregates all the way
-    to compute a constant upper bounds depending on the perturbation domain
-
-    Args:
-        x: the tensors that represent the domain
-        w: the weights of the affine function
-        b: the bias
-        perturbation_domain: the type of perturbation domain (see ???)
-
-    Returns:
-        a constant upper bound of the affine function
-    """
-
-    if perturbation_domain is None:
-        perturbation_domain = BoxDomain()
-
-    if isinstance(perturbation_domain, (BoxDomain, GridDomain)):
-        x_min = x[:, 0]
-        x_max = x[:, 1]
-        return get_upper_box(x_min, x_max, w, b, **kwargs)
-
-    elif isinstance(perturbation_domain, BallDomain):
-        eps = perturbation_domain.eps
-        p = perturbation_domain.p
-        return get_upper_ball(x, eps, p, w, b, **kwargs)
-
-    else:
-        raise NotImplementedError(f"Not implemented for perturbation domain type {type(perturbation_domain)}")
-
-
-def get_lower(
-    x: tf.Tensor, w: tf.Tensor, b: tf.Tensor, perturbation_domain: Optional[PerturbationDomain] = None, **kwargs: Any
-) -> tf.Tensor:
-    """Meta function that aggregates all the way
-    to compute a constant lower bound depending on the perturbation domain
-        :param x: the tensors that represent the domain
-        :param w: the weights of the affine function
-        :param b: the bias
-        :param perturbation_domain: the type of perturbation domain (see ???)
-        :return: a constant upper bound of the affine function
-    """
-    if perturbation_domain is None:
-        perturbation_domain = BoxDomain()
-
-    if isinstance(perturbation_domain, (BoxDomain, GridDomain)):
-        x_min = x[:, 0]
-        x_max = x[:, 1]
-        return get_lower_box(x_min, x_max, w, b, **kwargs)
-
-    elif isinstance(perturbation_domain, BallDomain):
-        eps = perturbation_domain.eps
-        p = perturbation_domain.p
-        return get_lower_ball(x, eps, p, w, b, **kwargs)
-
-    else:
-        raise NotImplementedError(f"Not implemented for perturbation domain type {type(perturbation_domain)}")
 
 
 def get_lower_layer(perturbation_domain: Optional[PerturbationDomain] = None) -> Layer:
