+++ tests/test_utils.py
@@ -6,8 +6,8 @@ import pytest
 import tensorflow.keras.backend as K
 from numpy.testing import assert_allclose, assert_almost_equal
 
-from decomon.core import ForwardMode
-from decomon.layers.utils import add, get_lower, get_upper, max_, maximum, minus, relu_
+from decomon.core import ForwardMode, get_lower, get_upper
+from decomon.layers.utils import add, max_, maximum, minus, relu_
 from decomon.utils import subtract
